Deleting request from queue after completion of task

+++ DrugInventoryEcoSystem/src/userinterface/manufacturer/producerrole/ProducerWorkAreaJPanel.java
@@ -81,6 +81,7 @@ public class ProducerWorkAreaJPanel extends javax.swing.JPanel {
         title = new javax.swing.JLabel();
         processOrderjButton = new javax.swing.JButton();
         completeOrderjButton = new javax.swing.JButton();
+        showAnalyticsJButton = new javax.swing.JButton();
 
         jPanel1.setPreferredSize(new java.awt.Dimension(1200, 750));
 
@@ -98,7 +99,7 @@ public class ProducerWorkAreaJPanel extends javax.swing.JPanel {
                 {null, null, null, null}
             },
             new String [] {
-                "Request ID", "Status", "Supplier", "Manufacturer"
+                "Request ID", "Status", "Sender", "Receiver"
             }
         ) {
             Class[] types = new Class [] {
@@ -135,7 +136,7 @@ public class ProducerWorkAreaJPanel extends javax.swing.JPanel {
         });
 
         title.setFont(new java.awt.Font("Segoe UI", 1, 24)); // NOI18N
-        title.setText("Approver Workarea");
+        title.setText("Producer Workarea");
         title.setAlignmentX(740.0F);
         title.setAlignmentY(245.0F);
 
@@ -153,6 +154,13 @@ public class ProducerWorkAreaJPanel extends javax.swing.JPanel {
             }
         });
 
+        showAnalyticsJButton.setText("Show Analytics");
+        showAnalyticsJButton.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                showAnalyticsJButtonActionPerformed(evt);
+            }
+        });
+
         javax.swing.GroupLayout kGradientPanel1Layout = new javax.swing.GroupLayout(kGradientPanel1);
         kGradientPanel1.setLayout(kGradientPanel1Layout);
         kGradientPanel1Layout.setHorizontalGroup(
@@ -167,9 +175,11 @@ public class ProducerWorkAreaJPanel extends javax.swing.JPanel {
                         .addGroup(kGradientPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
                             .addGroup(kGradientPanel1Layout.createSequentialGroup()
                                 .addComponent(assignToMe)
-                                .addGap(138, 138, 138)
-                                .addComponent(processOrderjButton)
+                                .addGap(18, 18, 18)
+                                .addComponent(showAnalyticsJButton)
                                 .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                                .addComponent(processOrderjButton)
+                                .addGap(18, 18, 18)
                                 .addComponent(completeOrderjButton))
                             .addComponent(jScrollPane1, javax.swing.GroupLayout.PREFERRED_SIZE, 669, javax.swing.GroupLayout.PREFERRED_SIZE))))
                 .addContainerGap(498, Short.MAX_VALUE))
@@ -185,7 +195,8 @@ public class ProducerWorkAreaJPanel extends javax.swing.JPanel {
                 .addGroup(kGradientPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
                     .addComponent(assignToMe)
                     .addComponent(processOrderjButton)
-                    .addComponent(completeOrderjButton))
+                    .addComponent(completeOrderjButton)
+                    .addComponent(showAnalyticsJButton))
                 .addContainerGap(386, Short.MAX_VALUE))
         );
 
@@ -279,7 +290,15 @@ public class ProducerWorkAreaJPanel extends javax.swing.JPanel {
             request.setSender(request.getReceiver());
             request.setReceiver(supplier);
             JOptionPane.showMessageDialog(null, "Order completed successfully");
-            organization.getWorkQueue().deleteWorkRequest(request);
+
+            //Delete this order from all queues.
+            for (Organization organization : enterprise.getOrganizationDirectory().getOrganizationList()) {
+                if (organization instanceof ProducerOrganization) {
+                    //Remove the workrequest from this queue
+                    organization.getWorkQueue().deleteWorkRequest(request);
+                }
+            }
+
             populateRequestTable();
         } else {
             JOptionPane.showMessageDialog(null, "Please process the order first before completing it");
@@ -288,6 +307,10 @@ public class ProducerWorkAreaJPanel extends javax.swing.JPanel {
 
     }//GEN-LAST:event_completeOrderjButtonActionPerformed
 
+    private void showAnalyticsJButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_showAnalyticsJButtonActionPerformed
+        // TODO add your handling code here:
+    }//GEN-LAST:event_showAnalyticsJButtonActionPerformed
+
     // Variables declaration - do not modify//GEN-BEGIN:variables
     private javax.swing.JButton assignToMe;
     private javax.swing.JButton completeOrderjButton;
@@ -295,6 +318,7 @@ public class ProducerWorkAreaJPanel extends javax.swing.JPanel {
     private javax.swing.JScrollPane jScrollPane1;
     private keeptoo.KGradientPanel kGradientPanel1;
     private javax.swing.JButton processOrderjButton;
+    private javax.swing.JButton showAnalyticsJButton;
     private javax.swing.JLabel title;
     private javax.swing.JTable workRequestJTable;
     // End of variables declaration//GEN-END:variables
